extract vindextable to it's own file
