+++ go/vt/vtgate/semantics/real_table.go
@@ -17,6 +17,8 @@ limitations under the License.
 package semantics
 
 import (
+	"strings"
+
 	vtrpcpb "vitess.io/vitess/go/vt/proto/vtrpc"
 	"vitess.io/vitess/go/vt/sqlparser"
 	"vitess.io/vitess/go/vt/vterrors"
@@ -35,7 +37,17 @@ var _ TableInfo = (*RealTable)(nil)
 
 // Dependencies implements the TableInfo interface
 func (r *RealTable) dependencies(colName string, org originable) (dependencies, error) {
-	return depsForAliasedAndRealTables(colName, org, r.ASTNode, r.getColumns(), r.authoritative())
+	ts := org.tableSetFor(r.ASTNode)
+	for _, info := range r.getColumns() {
+		if strings.EqualFold(info.Name, colName) {
+			return createCertain(ts, ts, &info.Type), nil
+		}
+	}
+
+	if r.authoritative() {
+		return &nothing{}, nil
+	}
+	return createUncertain(ts, ts), nil
 }
 
 // GetTables implements the TableInfo interface
@@ -81,4 +93,36 @@ func (r *RealTable) authoritative() bool {
 // Matches implements the TableInfo interface
 func (r *RealTable) matches(name sqlparser.TableName) bool {
 	return (name.Qualifier.IsEmpty() || name.Qualifier.String() == r.dbName) && r.tableName == name.Name.String()
+}
+
+func vindexTableToColumnInfo(tbl *vindexes.Table) []ColumnInfo {
+	if tbl == nil {
+		return nil
+	}
+	nameMap := map[string]interface{}{}
+	cols := make([]ColumnInfo, 0, len(tbl.Columns))
+	for _, col := range tbl.Columns {
+		cols = append(cols, ColumnInfo{
+			Name: col.Name.String(),
+			Type: col.Type,
+		})
+		nameMap[col.Name.String()] = nil
+	}
+	// If table is authoritative, we do not need ColumnVindexes to help in resolving the unqualified columns.
+	if tbl.ColumnListAuthoritative {
+		return cols
+	}
+	for _, vindex := range tbl.ColumnVindexes {
+		for _, column := range vindex.Columns {
+			name := column.String()
+			if _, exists := nameMap[name]; exists {
+				continue
+			}
+			cols = append(cols, ColumnInfo{
+				Name: name,
+			})
+			nameMap[name] = nil
+		}
+	}
+	return cols
 }
